Updated API for Product Management

See merge request dev-training/tghuy!6

+++ src/main/java/com/tma/coffeehouse/Product/ProductService.java
@@ -5,6 +5,7 @@ import com.tma.coffeehouse.Product.DTO.AddProductRequest;
 import com.tma.coffeehouse.Product.DTO.ProductDTO;
 import com.tma.coffeehouse.Product.DTOMapper.AddProductMapper;
 import com.tma.coffeehouse.Product.DTOMapper.ProductMapper;
+import com.tma.coffeehouse.ProductCategory.ProductCategoryService;
 import com.tma.coffeehouse.Topping.Topping;
 import com.tma.coffeehouse.Topping.ToppingRepository;
 import com.tma.coffeehouse.Topping.ToppingService;
@@ -14,6 +15,8 @@ import com.tma.coffeehouse.ProductCategory.ProductCategory;
 import com.tma.coffeehouse.ProductCategory.ProductCategoryRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -26,6 +29,8 @@ import org.springframework.web.multipart.MultipartFile;
 import java.io.File;
 
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.*;
 
 @Service
@@ -34,9 +39,23 @@ public class ProductService {
     private final ProductRepository productRepository;
     private final AddProductMapper addProductMapper;
     private final ProductMapper productMapper;
-    Page<Product> findAll(Long productCategoryID, String name, Integer pageNo, Integer pageSize, String sortBy){
-        Pageable pageAndSortingRequest = PageRequest.of(pageNo, pageSize, Sort.by(sortBy).ascending());
+    @Lazy
+    @Autowired
+    private ToppingService toppingService;
+    private final ProductCategoryService productCategoryService;
+    Page<Product> findAll(Long productCategoryID, String name, Integer pageNo, Integer pageSize, String sortBy, Boolean reverse){
+        Pageable pageAndSortingRequest = PageRequest.of(pageNo, pageSize, Sort.by(reverse? Sort.Direction.DESC : Sort.Direction. ASC, sortBy));
         return  productRepository.findAllByQueries(productCategoryID, name, pageAndSortingRequest);
+    }
+    byte[] getImage(Long id){
+        Product currentProd = productRepository.findById(id)
+                .orElseThrow(() -> new CustomException("Không tìm thấy sản phẩm với mã " + id, HttpStatus.NOT_FOUND));
+        try{
+            return Files.readAllBytes(Paths.get("./src/main/resources/static/prod-img/" + id + "/" +  currentProd.getImage()));
+        }
+        catch(IOException e){
+            throw new CustomException("Có lỗi xảy ra khi lấy hình ảnh sản phẩm", HttpStatus.NOT_FOUND);
+        }
     }
     ProductDTO findById(Long id){
         Product product = productRepository.findById(id).orElseThrow(()->new CustomException("Không tìm thấy sản phẩm với mã " + id, HttpStatus.NOT_FOUND));
@@ -74,11 +93,24 @@ public class ProductService {
     public ProductDTO update (Long id, AddProductRequest productRequest,  MultipartFile multipartFile){
         Product currentProd = productRepository.findById(id)
                 .orElseThrow(() -> new CustomException("Không tìm thấy sản phẩm với mã " + id, HttpStatus.NOT_FOUND));
-        CustomUtils.deleteFile("./src/main/resources/static/prod-img/" + currentProd.getId() + "/" + currentProd.getImage());
         currentProd.setName(productRequest.getName());
         currentProd.setPrice(productRequest.getPrice());
         currentProd.setDescription(productRequest.getDescription());
+        ProductCategory cate = productCategoryService.findById(productRequest.getProductCategoryId());
+        currentProd.setProductCategory(cate);
+        Set<Topping> toppingsOfProduct = toppingService.findManyByIDs(productRequest.getToppingsId());
+        currentProd.setProductToppings(toppingsOfProduct);
 
+        // If not having a new image from request to update, then return here.
+        if (multipartFile == null){
+            Product saved =  productRepository.save(currentProd);
+            for (Topping topping : toppingsOfProduct){
+                topping.addProduct(saved);
+            }
+            return productMapper.modelTODto(saved);
+        }
+
+        CustomUtils.deleteFile("./src/main/resources/static/prod-img/" + currentProd.getId() + "/" + currentProd.getImage());
         String fileName = StringUtils.cleanPath(Objects.requireNonNull(multipartFile.getOriginalFilename()));
         currentProd.setImage(fileName);
         String uploadDir = "./src/main/resources/static/prod-img/" + currentProd.getId();
